Allows saving edits, adds tags to edit view, moves TextEntryFormElement and ItemDetail to separate files
