+++ src/ItemTable.js
@@ -5,148 +5,20 @@
 var React = require('react');
 var ReactBsTable = require('react-bootstrap-table');
 var Bootstrap = require('react-bootstrap');
+import ItemDetail from './ItemDetail';
 var BootstrapTable = ReactBsTable.BootstrapTable;
 var TableHeaderColumn = ReactBsTable.TableHeaderColumn;
-var Modal = Bootstrap.Modal;
-var Button = Bootstrap.Button;
-var Form = Bootstrap.Form;
-var FormGroup = Bootstrap.FormGroup;
-var Col = Bootstrap.Col;
-var ControlLabel = Bootstrap.ControlLabel;
-var FormControl = Bootstrap.FormControl;
 
 import { hashHistory } from 'react-router';
 
 var xhttp = new XMLHttpRequest();
-
-class EditFormElement extends React.Component {
-  constructor(props) {
-    super(props);
-
-    this.state = {
-      value: (this.props.initialValue == null) ? "" : this.props.initialValue
-    }
-    this.handleChange = this.handleChange.bind(this);
-  }
-
-  handleChange(evt) {
-    this.setState({value: evt.target.value});
-  }
-
-  render() {
-    return (
-      <FormGroup controlId={this.props.controlId}>
-      <Col componentClass={ControlLabel} sm={2}>
-      {this.props.label}
-      </Col>
-      <Col sm={10}>
-      <FormControl componentClass={this.props.componentClass} type={this.props.type} value={this.state.value} onChange={this.handleChange}/>
-      </Col>
-      </FormGroup>
-    )
-  }
-}
-
-class ItemDetail extends React.Component {
-  constructor(props) {
-    super(props);
-    this.state = {
-      showModal: false,
-      isAdmin: true,
-      isEditing: false
-    }
-    this.openModal = this.openModal.bind(this);
-    this.closeModal = this.closeModal.bind(this);
-    this.makeRequest = this.makeRequest.bind(this);
-    this.toggleEditing = this.toggleEditing.bind(this);
-    this.saveEdits = this.saveEdits.bind(this);
-  }
-
-  openModal() {
-    this.setState({showModal: true});
-  }
-
-  closeModal() {
-    if(this.state.isEditing) {
-      this.toggleEditing();
-    }
-    this.setState({showModal: false});
-  }
-
-  makeRequest() {
-    //TODO: implement
-    console.log('request');
-  }
-
-  toggleEditing() {
-    this.setState({isEditing: !this.state.isEditing});
-  }
-
-  saveEdits() {
-    console.log("Save edits");
-  }
-
-  render() {
-    //TODO: Add in image
-    return (
-      <div>
-      <Bootstrap.Modal show={this.state.showModal}>
-      <Modal.Body>
-
-      {this.state.isEditing ?
-        <Form horizontal>
-        <EditFormElement controlId="formHorizontalName" label="Name" type="text" initialValue={this.props.name}/>
-        <EditFormElement controlId="formHorizontalQuantity" label="Quantity" type="number" initialValue={this.props.quantity}/>
-        <EditFormElement controlId="formHorizontalModelNumber" label="Model Number" type="number" initialValue={this.props.model_number}/>
-        <EditFormElement controlId="formHorizontalLocation" label="Location" type="text" initialValue={this.props.location}/>
-        <EditFormElement controlId="formHorizontalDescription" label="Description" type="text" initialValue={this.props.description} componentClass="textarea"/>
-        </Form>
-        :
-        <div>
-        <h2> {this.props.name} </h2>
-        <p> Quantity: {this.props.quantity} </p>
-        <p> Model Number: {this.props.model_number} </p>
-        <p> Location: {this.props.location} </p>
-        <p> Tags: {this.props.tags} </p>
-        <p> {this.props.description} </p>
-        </div>
-      }
-
-      </Modal.Body>
-      <Modal.Footer>
-      {this.state.isAdmin ?
-        this.state.isEditing ?
-        <div>
-        <Button onClick={this.saveEdits} bsStyle="primary">Save</Button>
-        <Button onClick={this.toggleEditing} bsStyle="danger">Cancel</Button>
-        </div>
-        :
-        <div>
-        <Button onClick={this.toggleEditing} bsStyle="primary">Edit</Button>
-        <Button onClick={this.closeModal} bsStyle="danger">Close</Button>
-        </div>
-        :
-        <Button onClick={this.closeModal} bsStyle="danger">Close</Button>
-      }
-
-      </Modal.Footer>
-      </Bootstrap.Modal>
-      </div>
-    )
-  }
-}
 
 class ItemTable extends React.Component {
 
   constructor(props) {
     super(props);
     this.state = {
-      name: null,
-      quantity: 0,
-      model_number: 0,
-      description: null,
-      location: null,
-      tags: null,
+      row: null,
       _products: [{
         "id": 111111111,
         "name": "siva",
@@ -221,9 +93,9 @@ class ItemTable extends React.Component {
     return returnList;
   }
 
-  onAddRow(row) {
+  addOrUpdateRow(row, requestType, itemID) {
     if (row){
-      xhttp.open("POST", "https://asap-test.colab.duke.edu/api/item/", false);
+      xhttp.open(requestType, "https://asap-test.colab.duke.edu/api/item/" + itemID, false);
       xhttp.setRequestHeader("Content-Type", "application/json");
       xhttp.setRequestHeader("Authorization", "Bearer " + localStorage.token);
       if (xhttp.status === 401 || xhttp.status === 500){
@@ -250,6 +122,10 @@ class ItemTable extends React.Component {
       }
     }
     console.log(this.state._products);
+  }
+
+  onAddRow(row) {
+    this.addOrUpdateRow(row, 'POST', '');
   }
 
   onDeleteRow(rows) {
@@ -285,12 +161,7 @@ class ItemTable extends React.Component {
   }
 
   onRowClick(row, isSelected, e) {
-    this.setState({name: row.name});
-    this.setState({quantity: row.quantity});
-    this.setState({model_number: row.model_number});
-    this.setState({description: row.description});
-    this.setState({location: row.location});
-    this.setState({tags: row.tags});
+    this.setState({row: row});
     this._child.openModal();
   }
 
@@ -321,10 +192,8 @@ class ItemTable extends React.Component {
       </BootstrapTable>) : null}
 
       <ItemDetail  ref={(child) => { this._child = child; }}
-        name={this.state.name} quantity={this.state.quantity} model_number={this.state.model_number}
-        description={this.state.description} location={this.state.location} tags={this.tagsToListString(this.state.tags)}/>
+      row={this.state.row} updateCallback={this}/>
       </div>
-
     )
   }
 }
