+++ src/TextEntryFormElement
@@ -1,0 +1,42 @@
+var React = require('react');
+var ReactBsTable = require('react-bootstrap-table');
+var Bootstrap = require('react-bootstrap');
+var BootstrapTable = ReactBsTable.BootstrapTable;
+var TableHeaderColumn = ReactBsTable.TableHeaderColumn;
+var Modal = Bootstrap.Modal;
+var Button = Bootstrap.Button;
+var Form = Bootstrap.Form;
+var FormGroup = Bootstrap.FormGroup;
+var Col = Bootstrap.Col;
+var ControlLabel = Bootstrap.ControlLabel;
+
+// Text entry element for React Bootstrap formHorizontalDescription
+// @author Patrick Terry
+
+class TextEntryFormElement extends React.Component {
+  constructor(props) {
+    super(props);
+
+    this.state = {
+      value: (this.props.initialValue == null) ? "" : this.props.initialValue
+    }
+    this.handleChange = this.handleChange.bind(this);
+  }
+
+  handleChange(evt) {
+    this.setState({value: evt.target.value});
+  }
+
+  render() {
+    return (
+      <FormGroup controlId={this.props.controlId}>
+      <Col componentClass={ControlLabel} sm={2}>
+      {this.props.label}
+      </Col>
+      <Col sm={10}>
+      <FormControl componentClass={this.props.componentClass} type={this.props.type} value={this.state.value} onChange={this.handleChange}/>
+      </Col>
+      </FormGroup>
+    )
+  }
+}
