+++ src/UserTabs.js
@@ -3,10 +3,8 @@
 // @author Andrew
 
 import ItemTable from './ItemTable'
-
 var React = require('react');
 var ReactBootstrap = require('react-bootstrap');
-
 var Tab = ReactBootstrap.Tab;
 var Row = ReactBootstrap.Row;
 var Col = ReactBootstrap.Col;
